Tiger level assignment. I don't know if writing the test the way I did was the best way, the test knows too much about the implementation of the code, which makes it brittle... which way would you recommend for this?
